requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pinecone-client==6.0.0
langchain-pinecone==0.2.12
langchain-openai==0.3.33
langchain-core==0.3.76
langchain==0.3.27
openai==1.107.2
pydantic==2.11.9
httpx==0.28.1
anyio==4.10.0